refactor(spice): move gc test-loop tests to garbage_collection.rs
